Refactor CLI into SRP modules
